Fix loading of globals from love2d generated documentation.
The first thing to change is to load the love2d docset inside a pcall,
because it's not included in the standard `make` build, so we need to
not crash if it's not present. Note that we are storing these under
libraries.love2d rather than libraries.love-2d, because even tho it's
inconsistent with tic-80; I've never seen them use the hyphenated
version, and now the name matches the module name.

Secondly, we extend the get-global function to work with anything in
the server.configuration.libraries table, not just hard-coded to
tic-80.

Finally, update the manual to reflect the removal of the love2d hypen.