Improve selflint results by fixing a few mistakes.
But selflint still does not pass. This appears to be due to bugs:

    src/fennel-ls/searcher.fnl:9:10: unknown field: path-sep
    src/fennel-ls.fnl:44:2: unknown field: table.unpack

The first one is a pretty clear issue with the field being exported
from fennel-ls.utils but somehow not picked up; reason is unclear.

The second is due to the macroexpansion of `case` making reference to
`table.unpack`, but I think it should be suppressed because the
unknown global is not in this code; it's in the macro.

Once these get fixed, we can add selflint to the ci target.