Load libraries docsets from disk at runtime instead of bundling them
The docsets are the same .lua files that are currently being generated
by `make docs`, renamed to .fls to distinguish them from regular code
files.

The files should be copied to $XDG_DATA_HOME/fennel-ls/docsets/, the
subdirectory allows the directory to be used for other things in the
future, maybe some global configutation.

The docset are only loaded if the librarys are added to flsproject.fnl
with the same syntax used until now.